+++ WHATSNEW.md
@@ -8,6 +8,7 @@
  + added openstreetmap.org plugin (Mathias Panzenböck)
  + added marketplace.firefox.com plugin (Mathias Panzenböck)
  * fixed smugmug.com plugin
+ + added pinterest.com pin-board widget plugin
 
 0.3.10